#149 Separate ekubo pools and ekubo fav pools
